drop uniter_old_test, move components into uniter_test and util_test as appropriate
